+++ module/render/src/commonTest/kotlin/org/umamo/render/puppet/GlueLayoutTest.kt
@@ -61,7 +61,7 @@ class GlueLayoutTest {
 			canvasWidth = 0f,
 			canvasHeight = 0f,
 			worldOriginX = 0f,
-			worldOriginY = 0f,
+			worldOriginZ = 0f,
 		)
 
 	@Test