+++ module/render/src/jvmTest/kotlin/org/umamo/render/gl/MaskCoverageBlendTest.kt
@@ -79,7 +79,7 @@ class MaskCoverageBlendTest {
 			canvasWidth = 0f,
 			canvasHeight = 0f,
 			worldOriginX = 0f,
-			worldOriginY = 0f,
+			worldOriginZ = 0f,
 		)
 	}